Moved ConnectGame class to connect_game.py
