+++ game.py
@@ -9,9 +9,9 @@ from pygame import mixer
 from pygame.gfxdraw import aacircle, filled_circle
 from pygame.locals import KEYDOWN
 
-from assets import *
+
 from config import *
-from config import black
+from connect_game import ConnectGame
 from game_data import GameData
 from graphics import GameRenderer
 
@@ -24,123 +24,6 @@ radius: int = int(sq_size / 2 - 5)
 
 pygame.init()
 screen = pygame.display.set_mode(size)
-
-class ConnectGame:
-    game_data: GameData
-    renderer: GameRenderer
-
-    def __init__(self, game_data: GameData, renderer: GameRenderer):
-        self.game_data = game_data
-        self.renderer = renderer
-
-    def quit(self):
-        sys.exit()
-
-    def mouse_move(self, posx: int):
-        # Queue message that mouse has been moved
-
-        ### Renderer should see action is mouse move, and draw this ##########################
-        pygame.draw.rect(self.renderer.screen, black, (0, 0, width, sq_size))
-        if self.game_data.turn == 0:
-            self.renderer.draw_red_coin(posx - (sq_size / 2), int(sq_size) - sq_size + 5)
-        else:
-            self.renderer.draw_yellow_coin(posx - (sq_size / 2), int(sq_size) - sq_size + 5)
-        #######################################################################################
-    def mouse_click(self, posx: int):
-        ### Renderer should see action is mouse click, and draw this ##########################
-        pygame.draw.rect(self.renderer.screen, black, (0, 0, width, sq_size))
-        #######################################################################################
-        if self.game_data.turn == 0:
-
-            col: int = int(math.floor(posx / sq_size))
-
-            if self.game_data.game_board.is_valid_location(col):
-                row: int = self.game_data.game_board.get_next_open_row( col)
-
-                self.game_data.last_move_row = row
-                self.game_data.last_move_col = col
-                self.game_data.game_board.drop_piece(row, col, 1)
-
-            self.print_board()
-
-            if self.game_data.game_board.winning_move(1):
-                ### Move to renderer #########################
-                self.renderer.label = self.renderer.myfont.render("PLAYER 1 WINS!", 1, red)
-                self.renderer.screen.blit(self.renderer.label, (40, 10))
-                ###############################################
-
-                mixer.music.load(event_sound)
-                mixer.music.play(0)
-                self.game_data.game_over = True
-            pygame.display.update()
-        else:
-            col: int = int(math.floor(posx / sq_size))
-
-            if self.game_data.game_board.is_valid_location(col):
-                row: int = self.game_data.game_board.get_next_open_row( col)
-
-                self.game_data.last_move_row = row
-                self.game_data.last_move_col = col
-                self.game_data.game_board.drop_piece( row, col, 2)
-
-            self.print_board()
-
-            if self.game_data.game_board.winning_move( 2):
-                ##### Move to renderer ###
-                self.renderer.label = self.renderer.myfont.render("PLAYER 2 WINS!", 1, yellow)
-                screen.blit(self.renderer.label, (40, 10))
-                ##########################
-
-                mixer.music.load(event_sound)
-                mixer.music.play(0)
-                self.game_data.game_over = True
-            pygame.display.update()
-        self.game_data.turn += 1
-        self.game_data.turn = self.game_data.turn % 2
-
-    def undo(self):
-        self.game_data.game_board.drop_piece(
-            self.game_data.last_move_row,
-            self.game_data.last_move_col,
-            0
-        )
-
-        ################## Move to Renderer #######################
-        filled_circle(
-            screen,
-            self.game_data.last_move_row,
-            self.game_data.last_move_col,
-            radius,
-            black
-        )
-
-        aacircle(
-            screen,
-            self.game_data.last_move_row,
-            self.game_data.last_move_col,
-            radius,
-            black
-        )
-
-        self.renderer.draw_black_coin(
-            self.game_data.last_move_col * sq_size + 5,
-            height - (self.game_data.last_move_row * sq_size + sq_size - 5)
-        )
-
-        self.print_board()
-        ############################################################
-        self.game_data.turn += 1
-        self.game_data.turn = self.game_data.turn % 2
-
-    def update(self):
-        pass
-
-    def draw(self):
-        self.renderer.draw(game.game_data)
-
-    def print_board(self):
-        self.game_data.game_board.print_board()
-
 
 game = ConnectGame(GameData(), GameRenderer(screen))
